# Conflicts:
#	src/test/java/cn/rzhd/wuye/PerfectInformationTest.java

+++ src/main/java/cn/rzhd/wuye/mapper/TypeDifferentiateMapper.java
@@ -18,5 +18,5 @@ public interface TypeDifferentiateMapper {
     
      void updateTypeDifferentiateById(@Param("typeDifferentiate")TypeDifferentiate typeDifferentiate);
      
-     TypeDifferentiate getDataName(Integer typeDifferentiateId);
+     TypeDifferentiate getDataName(Long typeDifferentiateId);
 }
